Split Greview command implementation
