+++ lua/diffs/commands.lua
@@ -8,6 +8,7 @@ local git = require('diffs.git')
 local hunk_model = require('diffs.hunks')
 local log = require('diffs.log')
 local render = require('diffs.render')
+local review = require('diffs.review')
 local runtime = require('diffs.runtime')
 
 local dbg = log.dbg
@@ -378,6 +379,29 @@ local function replace_combined_diffs(raw_lines, repo_root)
   return result
 end
 
+---@return diffs.ReviewDeps
+local function review_deps()
+  return {
+    create_generated_diff_buffer = create_generated_diff_buffer,
+    show_generated_diff_buffer = show_generated_diff_buffer,
+    attach_generated_diff_buffer = attach_generated_diff_buffer,
+    replace_combined_diffs = replace_combined_diffs,
+  }
+end
+
+---@param spec? diffs.GreviewSpec
+---@return integer?
+function M.greview(spec)
+  return review.greview(spec, review_deps())
+end
+
+---@param buf integer
+---@param lnum integer
+---@return string?
+function M.review_file_at_line(buf, lnum)
+  return review.file_at_line(buf, lnum)
+end
+
 ---@param args? string
 ---@param vertical? boolean
 function M.gdiff(args, vertical)
@@ -591,423 +615,6 @@ function M.gdiff_section(repo_root, opts)
   dbg('opened section diff buffer %d (%s)', diff_buf, diff_label)
 end
 
--- selene: allow(global_usage)
-function _G._diffs_qftf(info)
-  local items = info.quickfix == 1 and vim.fn.getqflist({ id = info.id, items = 0 }).items
-    or vim.fn.getloclist(0, { id = info.id, items = 0 }).items
-  local max_lnum = 0
-  for i = info.start_idx, info.end_idx do
-    local e = items[i]
-    if e.lnum > 0 then
-      max_lnum = math.max(max_lnum, #tostring(e.lnum))
-    end
-  end
-  local lnum_fmt = '%' .. math.max(max_lnum, 1) .. 'd'
-  local lines = {}
-  for i = info.start_idx, info.end_idx do
-    local e = items[i]
-    local text = e.text or ''
-    if max_lnum > 0 and e.lnum > 0 then
-      table.insert(lines, ('%s  %s'):format(lnum_fmt:format(e.lnum), text))
-    else
-      table.insert(lines, text)
-    end
-  end
-  return lines
-end
-
----@class diffs.GreviewSpec
----@field base? string
----@field target? string
----@field repo? string
----@field mode? string
----@field vertical? boolean
-
----@class diffs.NormalizedGreview
----@field base string
----@field target string?
----@field repo_root string
----@field mode string?
----@field vertical boolean
----@field display string
----@field exec_args string[]
-
----@param repo? string
----@return string?
-local function resolve_repo_root(repo)
-  if repo and repo ~= '' then
-    return git.get_repo_root(repo .. '/.')
-  end
-
-  local bufnr = vim.api.nvim_get_current_buf()
-  local filepath = vim.api.nvim_buf_get_name(bufnr)
-  local repo_root = git.get_repo_root(filepath ~= '' and filepath or nil)
-  if repo_root then
-    return repo_root
-  end
-
-  local cwd = vim.fn.getcwd()
-  return git.get_repo_root(cwd .. '/.')
-end
-
----@param repo_root string
----@return string?
-local function default_branch(repo_root)
-  local ref =
-    vim.fn.systemlist({ 'git', '-C', repo_root, 'symbolic-ref', 'refs/remotes/origin/HEAD' })
-  if vim.v.shell_error ~= 0 or not ref[1] or ref[1] == '' then
-    return nil
-  end
-  return ref[1]:gsub('^refs/remotes/', '')
-end
-
----@param arg? string
----@return diffs.GreviewSpec?, string?
-local function parse_review_arg(arg)
-  if not arg or arg == '' then
-    return {}, nil
-  end
-
-  local base, target = arg:match('^(.-)%.%.%.(.+)$')
-  if base then
-    if base == '' or target == '' then
-      return nil, 'invalid review spec'
-    end
-    return { base = base, target = target, mode = 'merge-base' }, nil
-  end
-
-  base, target = arg:match('^(.-)%.%.(.+)$')
-  if base then
-    if base == '' or target == '' then
-      return nil, 'invalid review spec'
-    end
-    return { base = base, target = target, mode = 'direct' }, nil
-  end
-
-  return { base = arg }, nil
-end
-
----@param spec? diffs.GreviewSpec
----@param repo_root_override? string
----@return diffs.NormalizedGreview?, string?
-local function normalize_greview(spec, repo_root_override)
-  spec = spec or {}
-  if type(spec) ~= 'table' then
-    error('diffs: greview() expects a table spec')
-  end
-  if spec.base ~= nil and type(spec.base) ~= 'string' then
-    error('diffs: greview.base must be a string')
-  end
-  if spec.target ~= nil and type(spec.target) ~= 'string' then
-    error('diffs: greview.target must be a string')
-  end
-  if spec.repo ~= nil and type(spec.repo) ~= 'string' then
-    error('diffs: greview.repo must be a string')
-  end
-  if spec.mode ~= nil and type(spec.mode) ~= 'string' then
-    error('diffs: greview.mode must be a string')
-  end
-  if spec.vertical ~= nil and type(spec.vertical) ~= 'boolean' then
-    error('diffs: greview.vertical must be a boolean')
-  end
-
-  local repo_root = repo_root_override or resolve_repo_root(spec.repo)
-  if not repo_root then
-    return nil, 'not in a git repository'
-  end
-
-  local base = spec.base
-  if not base or base == '' then
-    base = default_branch(repo_root)
-    if not base then
-      return nil, 'cannot detect default branch (try: git remote set-head origin -a)'
-    end
-  end
-
-  local target = spec.target
-  if target == '' then
-    error('diffs: greview.target must be a non-empty string')
-  end
-
-  local mode = spec.mode
-  if target then
-    if mode == nil then
-      mode = 'merge-base'
-    elseif mode ~= 'merge-base' and mode ~= 'direct' then
-      error('diffs: greview.mode must be "merge-base" or "direct"')
-    end
-  elseif mode ~= nil then
-    error('diffs: greview.mode requires greview.target')
-  end
-
-  local display = base
-  local exec_args = { base }
-  if target then
-    if mode == 'merge-base' then
-      display = base .. '...' .. target
-      exec_args = { '--merge-base', base, target }
-    else
-      display = base .. '..' .. target
-      exec_args = { base, target }
-    end
-  end
-
-  return {
-    base = base,
-    target = target,
-    repo_root = repo_root,
-    mode = mode,
-    vertical = spec.vertical or false,
-    display = display,
-    exec_args = exec_args,
-  },
-    nil
-end
-
----@param review diffs.NormalizedGreview
----@return string[]
-local function build_review_cmd(review)
-  local cmd = { 'git', '-C', review.repo_root, 'diff', '--no-ext-diff', '--no-color' }
-  vim.list_extend(cmd, review.exec_args)
-  return cmd
-end
-
----@param review diffs.NormalizedGreview
----@return string
-local function review_buffer_name(review)
-  return 'diffs://review:' .. review.display
-end
-
----@param repo_root? string
----@return string[]
-local function list_refs(repo_root)
-  local cmd = { 'git' }
-  if repo_root then
-    table.insert(cmd, '-C')
-    table.insert(cmd, repo_root)
-  end
-  vim.list_extend(cmd, {
-    'for-each-ref',
-    '--format=%(refname:short)',
-    'refs/heads/',
-    'refs/remotes/',
-    'refs/tags/',
-  })
-  local refs = vim.fn.systemlist(cmd)
-  if vim.v.shell_error ~= 0 then
-    return {}
-  end
-  return refs
-end
-
----@param arglead string
----@return string[]
-local function complete_greview(arglead)
-  local refs = list_refs(resolve_repo_root(nil))
-  if arglead == '' then
-    return refs
-  end
-
-  local base, tail = arglead:match('^(.-)%.%.%.(.*)$')
-  if base then
-    if base == '' then
-      return {}
-    end
-    local matches = {}
-    for _, ref in ipairs(refs) do
-      if ref:find(tail, 1, true) == 1 then
-        table.insert(matches, base .. '...' .. ref)
-      end
-    end
-    return matches
-  end
-
-  base, tail = arglead:match('^(.-)%.%.(.*)$')
-  if base then
-    if base == '' then
-      return {}
-    end
-    local matches = {}
-    for _, ref in ipairs(refs) do
-      if ref:find(tail, 1, true) == 1 then
-        table.insert(matches, base .. '..' .. ref)
-      end
-    end
-    return matches
-  end
-
-  local matches = {}
-  for _, ref in ipairs(refs) do
-    if ref:find(arglead, 1, true) == 1 then
-      table.insert(matches, ref)
-    end
-  end
-  return matches
-end
-
----@param review diffs.NormalizedGreview
----@return string[]?, string?
-local function run_review(review)
-  local result = vim.fn.systemlist(build_review_cmd(review))
-  if vim.v.shell_error ~= 0 then
-    return nil, 'git diff failed'
-  end
-  return replace_combined_diffs(result, review.repo_root), nil
-end
-
----@param spec? diffs.GreviewSpec
----@return integer?
-function M.greview(spec)
-  local review, err = normalize_greview(spec)
-  if not review then
-    notify(err, vim.log.levels.ERROR)
-    return nil
-  end
-
-  local target_name = review_buffer_name(review)
-  local existing_buf = vim.fn.bufnr(target_name)
-  if existing_buf ~= -1 then
-    pcall(vim.api.nvim_buf_delete, existing_buf, { force = true })
-  end
-
-  local result, diff_err = run_review(review)
-  if not result then
-    notify(diff_err, vim.log.levels.ERROR)
-    return nil
-  end
-  if #result == 0 then
-    notify('no diff against ' .. review.display, vim.log.levels.INFO)
-    return nil
-  end
-
-  local diff_buf = create_generated_diff_buffer({
-    name = target_name,
-    lines = result,
-    repo_root = review.repo_root,
-    vars = {
-      diffs_review_base = review.base,
-      diffs_review_target = review.target,
-      diffs_review_mode = review.mode,
-    },
-  })
-
-  local qf_items = {}
-  local loc_items = {}
-  local current_file = nil
-  local file_adds, file_dels = {}, {}
-  local file_hunk_count = {}
-
-  for i, line in ipairs(result) do
-    local file = line:match('^diff %-%-git a/.+ b/(.+)$')
-    if file then
-      current_file = file
-      file_adds[file] = 0
-      file_dels[file] = 0
-      file_hunk_count[file] = 0
-      table.insert(qf_items, {
-        bufnr = diff_buf,
-        lnum = i,
-        text = file,
-      })
-    elseif current_file and line:match('^@@') then
-      file_hunk_count[current_file] = file_hunk_count[current_file] + 1
-      table.insert(loc_items, {
-        bufnr = diff_buf,
-        lnum = i,
-        text = current_file,
-        _hunk = file_hunk_count[current_file],
-        _header = line:match('^(@@.-@@)') or '',
-      })
-    elseif current_file then
-      local ch = line:sub(1, 1)
-      if ch == '+' and not line:match('^%+%+%+') then
-        file_adds[current_file] = file_adds[current_file] + 1
-      elseif ch == '-' and not line:match('^%-%-%-') then
-        file_dels[current_file] = file_dels[current_file] + 1
-      end
-    end
-  end
-
-  local max_fname = 0
-  local max_add, max_del = 0, 0
-  for _, item in ipairs(qf_items) do
-    max_fname = math.max(max_fname, #item.text)
-    local a = file_adds[item.text] or 0
-    local d = file_dels[item.text] or 0
-    if a > 0 then
-      max_add = math.max(max_add, #tostring(a) + 1)
-    end
-    if d > 0 then
-      max_del = math.max(max_del, #tostring(d) + 1)
-    end
-  end
-
-  for _, item in ipairs(qf_items) do
-    local file = item.text
-    local a = file_adds[file] or 0
-    local d = file_dels[file] or 0
-    local padded = file .. string.rep(' ', max_fname - #file)
-    local parts = { padded }
-    if max_add > 0 then
-      parts[#parts + 1] = a > 0 and string.format('%' .. max_add .. 's', '+' .. a)
-        or string.rep(' ', max_add)
-    end
-    if max_del > 0 then
-      parts[#parts + 1] = d > 0 and string.format('%' .. max_del .. 's', '-' .. d)
-        or string.rep(' ', max_del)
-    end
-    item.text = table.concat(parts, '  '):gsub('%s+$', '')
-  end
-
-  local max_loc_fname = 0
-  for _, item in ipairs(loc_items) do
-    max_loc_fname = math.max(max_loc_fname, #item.text)
-  end
-  for _, item in ipairs(loc_items) do
-    item.text = item.text
-      .. string.rep(' ', max_loc_fname - #item.text)
-      .. '  (hunk '
-      .. item._hunk
-      .. ') '
-      .. item._header
-    item._hunk = nil
-    item._header = nil
-  end
-
-  vim.fn.setqflist({}, ' ', {
-    title = 'review: ' .. review.display,
-    items = qf_items,
-    quickfixtextfunc = 'v:lua._diffs_qftf',
-  })
-
-  show_generated_diff_buffer(diff_buf, review.vertical)
-
-  vim.fn.setloclist(0, {}, ' ', {
-    title = 'review hunks: ' .. review.display,
-    items = loc_items,
-    quickfixtextfunc = 'v:lua._diffs_qftf',
-  })
-
-  attach_generated_diff_buffer(diff_buf)
-  dbg('opened review buffer %d (%s)', diff_buf, review.display)
-
-  return diff_buf
-end
-
----@param buf integer
----@param lnum integer
----@return string?
-function M.review_file_at_line(buf, lnum)
-  local lines = vim.api.nvim_buf_get_lines(buf, 0, lnum, false)
-  for i = #lines, 1, -1 do
-    local file = lines[i]:match('^diff %-%-git a/.+ b/(.+)$')
-    if file then
-      return file
-    end
-  end
-  return nil
-end
-
 ---@param bufnr integer
 function M.read_buffer(bufnr)
   local name = vim.api.nvim_buf_get_name(bufnr)
@@ -1057,36 +664,7 @@ function M.read_buffer(bufnr)
 
     diff_lines = replace_combined_diffs(diff_lines, repo_root)
   elseif not stored_spec and label == 'review' then
-    local stored_base = get_buf_var(bufnr, 'diffs_review_base')
-    local stored_target = get_buf_var(bufnr, 'diffs_review_target')
-    local stored_mode = get_buf_var(bufnr, 'diffs_review_mode')
-
-    local review_spec
-    if stored_base then
-      review_spec = {
-        base = stored_base,
-        target = stored_target,
-        mode = stored_mode,
-      }
-    else
-      review_spec = select(1, parse_review_arg(path))
-    end
-
-    if review_spec then
-      local review = select(1, normalize_greview(review_spec, repo_root))
-      if review then
-        diff_lines = vim.fn.systemlist(build_review_cmd(review))
-        if vim.v.shell_error ~= 0 then
-          diff_lines = {}
-        else
-          diff_lines = replace_combined_diffs(diff_lines, repo_root)
-        end
-      else
-        diff_lines = {}
-      end
-    else
-      diff_lines = {}
-    end
+    diff_lines = review.reload(bufnr, repo_root, path, review_deps())
   elseif not stored_spec then
     local abs_path = repo_root .. '/' .. path
 
@@ -1152,7 +730,7 @@ function M.setup()
   })
 
   vim.api.nvim_create_user_command('Greview', function(opts)
-    local spec, err = parse_review_arg(opts.args ~= '' and opts.args or nil)
+    local spec, err = review.parse_arg(opts.args ~= '' and opts.args or nil)
     if not spec then
       notify(err, vim.log.levels.ERROR)
       return
@@ -1160,16 +738,16 @@ function M.setup()
     M.greview(spec)
   end, {
     nargs = '?',
-    complete = complete_greview,
+    complete = review.complete,
     desc = 'Review the repo against the default branch or a git review spec',
   })
 end
 
 M._test = {
-  complete_greview = complete_greview,
+  complete_greview = review.complete,
   gdiff_buffer_label = gdiff_buffer_label,
-  normalize_greview = normalize_greview,
-  parse_review_arg = parse_review_arg,
+  normalize_greview = review.normalize,
+  parse_review_arg = review.parse_arg,
 }
 
 return M
